+++ lib/web_ui/lib/src/engine/compositor/canvas.dart
@@ -4,68 +4,39 @@
 
 part of engine;
 
-/// An actual [SkLayerCanvas] which can receive raw drawing commands.
-///
-/// In order for the drawing commands to be flushed to the associated HTML
-/// canvas, you must call `flush()` on the canvas's `SkSurface`.
-///
-/// Although this class is backed by an `SkCanvas` and can in theory perform
-/// arbitrary drawing operations, this class is only used in the final
-/// compositing by the layers, and arbitrary drawings are done in a
-/// [ui.Picture] which uses a Skia recording canvas. This class receives
-/// drawing calls from the various `Layer` classes, e.g. [ClipRectLayer] and
-/// so only exposes a subset of the drawing operations that can be performed
-/// on a canvas.
-class SkLayerCanvas {
+/// A wrapper around Skia's SKCanvas.
+class SkCanvas {
   final js.JsObject skCanvas;
-  final html.CanvasElement htmlCanvas;
-  final js.JsObject skSurface;
-  final ui.Size size;
 
-  SkLayerCanvas(this.skCanvas, this.htmlCanvas, this.skSurface, this.size);
+  SkCanvas(this.skCanvas);
 
-  int save() {
-    return skCanvas.callMethod('save');
+  @override
+  bool _didDraw = true;
+
+  @override
+  bool _hasArbitraryPaint = true;
+
+  @override
+  int get saveCount => skCanvas.callMethod('getSaveCount');
+
+  // This is required to implement RecordingCanvas.
+  @override
+  int _saveCount = -1;
+
+  @override
+  // TODO: implement _commands
+  List<PaintCommand> get _commands => null;
+
+  @override
+  // TODO: implement _paintBounds
+  _PaintBounds get _paintBounds => null;
+
+  @override
+  void apply(EngineCanvas engineCanvas) {
+    throw UnimplementedError("The Skia backend doesn't support apply()");
   }
 
-  int saveLayer(ui.Rect bounds, ui.Paint paint) {
-    return skCanvas.callMethod(
-        'saveLayer', <js.JsObject>[makeSkRect(bounds), makeSkPaint(paint)]);
-  }
-
-  int saveLayerWithFilter(ui.Rect bounds, ui.ImageFilter filter) {
-    final SkImageFilter skImageFilter = filter;
-    return skCanvas.callMethod(
-      'saveLayer',
-      <dynamic>[
-        null,
-        skImageFilter.skImageFilter,
-        0,
-        makeSkRect(bounds),
-      ],
-    );
-  }
-
-  void restore() {
-    skCanvas.callMethod('restore');
-  }
-
-  void restoreToCount(int count) {
-    skCanvas.callMethod('restoreToCount', <int>[count]);
-  }
-
-  void clear() {
-    skCanvas.callMethod('clear', <int>[0xffffffff]);
-  }
-
-  void translate(double dx, double dy) {
-    skCanvas.callMethod('translate', <double>[dx, dy]);
-  }
-
-  void transform(Float64List matrix) {
-    skCanvas.callMethod('concat', <js.JsArray<double>>[makeSkMatrix(matrix)]);
-  }
-
+  @override
   void clipPath(ui.Path path, {bool doAntiAlias = true}) {
     final SkPath skPath = path;
     final js.JsObject intersectClipOp = canvasKit['ClipOp']['Intersect'];
@@ -76,39 +47,255 @@ class SkLayerCanvas {
     ]);
   }
 
-  void clipRect(ui.Rect rect, {bool doAntiAlias = true}) {
+  @override
+  void clipRRect(
+    ui.RRect rrect, {
+    bool doAntiAlias = true,
+  }) {
     final js.JsObject intersectClipOp = canvasKit['ClipOp']['Intersect'];
-    skCanvas.callMethod('clipRect', <dynamic>[
-      makeSkRect(rect),
+    skCanvas.callMethod('clipRRect', <dynamic>[
+      makeSkRRect(rrect),
       intersectClipOp,
       doAntiAlias,
     ]);
   }
 
-  void clipRRect(ui.RRect rrect) {
-    final SkPath skPath = SkPath();
-    skPath.addRRect(rrect);
-    clipPath(skPath);
-  }
+  @override
+  void clipRect(
+    ui.Rect rect, {
+    ui.ClipOp clipOp = ui.ClipOp.intersect,
+    bool doAntiAlias = true,
+  }) {
+    js.JsObject skClipOp;
+    switch (clipOp) {
+      case ui.ClipOp.difference:
+        skClipOp = canvasKit['ClipOp']['Difference'];
+        break;
+      case ui.ClipOp.intersect:
+        skClipOp = canvasKit['ClipOp']['Intersect'];
+        break;
+    }
 
-  void drawPicture(ui.Picture picture) {
-    final SkPicture skPicture = picture;
-    skCanvas.callMethod('drawPicture', <js.JsObject>[skPicture.skPicture]);
-  }
-
-  void drawPath(ui.Path path, ui.Paint paint) {
-    final SkPath skPath = path;
     skCanvas.callMethod(
-        'drawPath', <js.JsObject>[skPath._skPath, makeSkPaint(paint)]);
+        'clipRect', <dynamic>[makeSkRect(rect), skClipOp, doAntiAlias]);
   }
 
+  @override
+  ui.Rect computePaintBounds() {
+    throw UnimplementedError(
+        "The Skia backend doesn't use computePaintBounds()");
+  }
+
+  @override
+  void debugDumpCommands() {
+    throw UnimplementedError(
+        "The Skia backend doesn't use debugDumpCommands()");
+  }
+
+  @override
+  void debugEnforceArbitraryPaint() {
+    throw UnimplementedError(
+        "The Skia backend doesn't use debugEnforceArbitraryPaint()");
+  }
+
+  @override
+  String debugPrintCommands() {
+    throw UnimplementedError(
+        "The Skia backend doesn't use debugPrintCommands()");
+  }
+
+  @override
+  bool get didDraw => true;
+
+  @override
+  void drawCircle(ui.Offset c, double radius, ui.Paint paint) {
+    skCanvas.callMethod('drawCircle', <dynamic>[
+      c.dx,
+      c.dy,
+      radius,
+      makeSkPaint(paint),
+    ]);
+  }
+
+  @override
+  void drawColor(ui.Color color, ui.BlendMode blendMode) {
+    skCanvas.callMethod('drawColor', <dynamic>[
+      color.value,
+      makeSkBlendMode(blendMode),
+    ]);
+  }
+
+  @override
+  void drawDRRect(ui.RRect outer, ui.RRect inner, ui.Paint paint) {
+    skCanvas.callMethod('drawDRRect', <js.JsObject>[
+      makeSkRRect(outer),
+      makeSkRRect(inner),
+      makeSkPaint(paint),
+    ]);
+  }
+
+  @override
+  void drawImage(ui.Image image, ui.Offset offset, ui.Paint paint) {
+    final SkImage skImage = image;
+    skCanvas.callMethod('drawImage', <dynamic>[
+      skImage.skImage,
+      offset.dx,
+      offset.dy,
+      makeSkPaint(paint),
+    ]);
+  }
+
+  @override
+  void drawImageRect(ui.Image image, ui.Rect src, ui.Rect dst, ui.Paint paint) {
+    final SkImage skImage = image;
+    skCanvas.callMethod('drawImageRect', <dynamic>[
+      skImage.skImage,
+      makeSkRect(src),
+      makeSkRect(dst),
+      makeSkPaint(paint),
+      false,
+    ]);
+  }
+
+  @override
+  void drawLine(ui.Offset p1, ui.Offset p2, ui.Paint paint) {
+    skCanvas.callMethod('drawLine', <dynamic>[
+      p1.dx,
+      p1.dy,
+      p2.dx,
+      p2.dy,
+      makeSkPaint(paint),
+    ]);
+  }
+
+  @override
+  void drawOval(ui.Rect rect, ui.Paint paint) {
+    skCanvas.callMethod('drawOval', <js.JsObject>[
+      makeSkRect(rect),
+      makeSkPaint(paint),
+    ]);
+  }
+
+  @override
   void drawPaint(ui.Paint paint) {
     skCanvas.callMethod('drawPaint', <js.JsObject>[makeSkPaint(paint)]);
   }
 
+  @override
+  void drawParagraph(ui.Paragraph paragraph, ui.Offset offset) {
+    // TODO(het): This doesn't support most paragraph features. We are just
+    // creating a font from the family and size, and drawing it with
+    // ShapedText.
+    final EngineParagraph engineParagraph = paragraph;
+    final ParagraphGeometricStyle style = engineParagraph.geometricStyle;
+    final js.JsObject skFont = skiaFontCollection.getFont(
+        style.effectiveFontFamily, style.fontSize ?? 12.0);
+    final js.JsObject skShapedTextOpts = js.JsObject.jsify(<String, dynamic>{
+      'font': skFont,
+      'leftToRight': true,
+      'text': engineParagraph.plainText,
+      'width': engineParagraph.width + 1,
+    });
+    final js.JsObject skShapedText =
+        js.JsObject(canvasKit['ShapedText'], <js.JsObject>[skShapedTextOpts]);
+    skCanvas.callMethod('drawText', <dynamic>[
+      skShapedText,
+      offset.dx + engineParagraph._alignOffset,
+      offset.dy,
+      makeSkPaint(engineParagraph._paint)
+    ]);
+  }
+
+  @override
+  void drawPath(ui.Path path, ui.Paint paint) {
+    final js.JsObject skPaint = makeSkPaint(paint);
+    final SkPath enginePath = path;
+    final js.JsObject skPath = enginePath._skPath;
+    skCanvas.callMethod('drawPath', <js.JsObject>[skPath, skPaint]);
+  }
+
+  @override
+  void drawRRect(ui.RRect rrect, ui.Paint paint) {
+    skCanvas.callMethod('drawRRect', <js.JsObject>[
+      makeSkRRect(rrect),
+      makeSkPaint(paint),
+    ]);
+  }
+
+  @override
+  void drawRect(ui.Rect rect, ui.Paint paint) {
+    final js.JsObject skRect = makeSkRect(rect);
+    final js.JsObject skPaint = makeSkPaint(paint);
+    skCanvas.callMethod('drawRect', <js.JsObject>[skRect, skPaint]);
+  }
+
+  @override
   void drawShadow(ui.Path path, ui.Color color, double elevation,
       bool transparentOccluder) {
     drawSkShadow(skCanvas, path, color, elevation, transparentOccluder,
         ui.window.devicePixelRatio);
+  }
+
+  @override
+  void drawVertices(
+      ui.Vertices vertices, ui.BlendMode blendMode, ui.Paint paint) {
+    SkVertices skVertices = vertices;
+    skCanvas.callMethod('drawVertices', <js.JsObject>[
+      skVertices.skVertices,
+      makeSkBlendMode(blendMode),
+      makeSkPaint(paint)
+    ]);
+  }
+
+  @override
+  bool get hasArbitraryPaint => true;
+
+  @override
+  void restore() {
+    skCanvas.callMethod('restore');
+  }
+
+  @override
+  void rotate(double radians) {
+    skCanvas
+        .callMethod('rotate', <double>[radians * 180.0 / math.pi, 0.0, 0.0]);
+  }
+
+  @override
+  void save() {
+    skCanvas.callMethod('save');
+  }
+
+  @override
+  void saveLayer(ui.Rect bounds, ui.Paint paint) {
+    skCanvas.callMethod('saveLayer', <js.JsObject>[
+      makeSkRect(bounds),
+      makeSkPaint(paint),
+    ]);
+  }
+
+  @override
+  void saveLayerWithoutBounds(ui.Paint paint) {
+    skCanvas.callMethod('saveLayer', <js.JsObject>[null, makeSkPaint(paint)]);
+  }
+
+  @override
+  void scale(double sx, double sy) {
+    skCanvas.callMethod('scale', <double>[sx, sy]);
+  }
+
+  @override
+  void skew(double sx, double sy) {
+    skCanvas.callMethod('skew', <double>[sx, sy]);
+  }
+
+  @override
+  void transform(Float64List matrix4) {
+    skCanvas.callMethod('concat', <js.JsArray<double>>[makeSkMatrix(matrix4)]);
+  }
+
+  @override
+  void translate(double dx, double dy) {
+    skCanvas.callMethod('translate', <double>[dx, dy]);
   }
 }
